Move Contents.js to Contents/index.js for consistency
